requirements.txt
twitchio==2.8.2
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.3
